refactor: 코드 전체 리팩토링
